+++ django_project_491/django_app/urls.py
@@ -15,4 +15,5 @@ urlpatterns = [
     path('create_question/', create_question, name='create_question'),
     path('list_questions/', list_questions_by_language, name='list_questions'),
     path('run_code/', run_code_view, name='run_code'),
+    path('get_api_languages/', get_run_coder_api_languages, name='get_run_coder_api_languages'),
 ]
